[Fix] remove `tests/quantization` (#99)
* initial commit

* woops

---------
