--- tests/quantization/test_quant_config.py
@@ -1,32 +1,0 @@
-# Copyright (c) 2021 - present / Neuralmagic, Inc. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import pytest
-from compressed_tensors.quantization import QuantizationConfig, QuantizationScheme
-
-
-@pytest.mark.parametrize(
-    "scheme_name",
-    [
-        "W8A8",
-        "W4A16",
-    ],
-)
-def test_load_scheme_from_preset(scheme_name: str):
-    targets = ["Linear"]
-    config = QuantizationConfig(config_groups={scheme_name: targets})
-
-    assert scheme_name in config.config_groups
-    assert isinstance(config.config_groups[scheme_name], QuantizationScheme)
-    assert config.config_groups[scheme_name].targets == targets
